refactor: Move OS detection functions to platform_utils.py and remove unused mica_transparency.py
