+++ monitor-app/main.py
@@ -7,7 +7,7 @@ from src.core.state import GameState, GameServerState
 from src.gui.utils.gui_init import init_gui
 from src.gui.utils.app_settings import AppSettings
 from src.core.app_settings import app_settings
-from PySide6.QtCore import QOperatingSystemVersion
+from src.gui.utils.platform_utils import is_windows_11
 
 gui_available = False
 
@@ -27,8 +27,6 @@ except Exception as e:
 if not gui_available:
     logger.warning("GUI functionality is disabled.")
 
-def is_windows_11():
-    return QOperatingSystemVersion.current() == QOperatingSystemVersion.Windows11
 
 def check_and_update_status(game_state, game_server_state, window):
     try:
