+++ monitor-app/src/gui/utils/gui_init.py
@@ -4,7 +4,8 @@ from src.gui.mainwindow import MainWindow
 from src.gui.system_tray import SystemTrayIcon
 from src.core.logger import logger
 import resources.resources # noqa: F401
-from src.gui.utils.gui_utils import get_current_theme, is_windows_11
+from src.gui.utils.gui_utils import get_current_theme
+from src.gui.utils.platform_utils import is_windows_11
 
 def init_gui():
     # import os
