+++ monitor-app/src/gui/utils/gui_utils.py
@@ -1,8 +1,8 @@
-import platform
 from PySide6.QtWidgets import QApplication
 from PySide6.QtGui import QPalette
 from src.gui.utils.gui_config import gui_config
-from PySide6.QtCore import QOperatingSystemVersion
+from src.gui.utils.platform_utils import is_linux
+
 
 def get_current_theme():
     app = QApplication.instance()
@@ -12,13 +12,10 @@ def get_current_theme():
     return 'light'  # Default to light theme if app instance is not available
 
 def get_icon_path(theme: str) -> str:
-    if platform.system() == 'Linux':
+    if is_linux():
         return f"{gui_config.ICON_THEME_DARK}/{gui_config.ICON_NAME}"
     icon_theme = gui_config.ICON_THEME_LIGHT if theme == 'light' else gui_config.ICON_THEME_DARK
     return f"{icon_theme}/{gui_config.ICON_NAME}"
-
-def is_linux():
-    return platform.system() == 'Linux'
 
 def apply_mica_transparency(window):
     window.setAttribute(Qt.WA_TranslucentBackground)
